store: move backup functions inside pkg/backup
